requirements.txt
numpy
annotated-types==0.7.0
click==8.2.1
pydantic==2.11.7
pydantic-core==2.33.2
python-dotenv==1.1.1
PyYAML==6.0.2
streamlit==1.40.1
streamlit-option-menu==0.4.0
pymongo==4.6.0
langchain==0.3.26
langchain-core==0.3.68
langchain-openai==0.3.27
langchain-text-splitters==0.3.8
langgraph==0.5.2
langgraph-checkpoint==2.1.0
langsmith==0.4.5
openai==1.95.0
requests==2.32.4
python-dateutil==2.8.2
altair==5.0.1
pandas
tavily-python==0.3.4
python-docx==0.8.11
